Translates the text.  Separates the pages.
Translates the text and stores them as a JSON file. Gets the amount of characters the can be displayed on the tablet, by using an instance of the tablet class. Separates the pages according to that amount. If the last character splits the word, finds the closest previous whitespace and separates the page from there. Defines turn_on, next, previous, and turn off function which are self descriptive. When called these function make sounds so that the reader knows which one was clicked.